Test skeletons for read/write requests
