+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/partition_ut/ya.make
@@ -1,9 +1,18 @@
 UNITTEST_FOR(ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct)
 
+FORK_SUBTESTS()
+
+REQUIREMENTS(ram:32 cpu:2)
+
+IF (SANITIZER_TYPE)
+    SIZE(LARGE)
+    INCLUDE(${ARCADIA_ROOT}/ydb/tests/large.inc)
+ELSE()
+    SIZE(MEDIUM)
+ENDIF()
+
 SRCS(
-    base_test_fixture.cpp
-    ddisk_data_copier_ut.cpp
-    read_request_ut.cpp
+    partition_direct_ut.cpp
 )
 
 PEERDIR(
